some modifications to scatter api to allow obtaining local edge/node counts

+++ include/graph/append_projection.hh
@@ -21,7 +21,7 @@ namespace neuroh5
      const NODE_IDX_T&         src_end,
      const NODE_IDX_T&         dst_start,
      const NODE_IDX_T&         dst_end,
-     const uint64_t&           num_edges,
+     const size_t&             num_edges,
      const edge_map_t&         prj_edge_map,
      const std::vector<std::vector<std::string>>& edge_attr_names,
      const hsize_t            cdim = 4096,
